Seperated out the main loop into a main loop class
