+++ 3D Graph/MainLoop.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "render.h"
+#include "Camera.h"
+#include <conio.h>
+
+class MainLoop
+{
+public:
+	static void loop(Screen& s, Camera& ca);
+	//static void UserInput();
+};
